#14 set note key when lazy loading children
the onLazyLoad method is executed when a node is expanded. there the
children for that node gets loaded. we set the key of the node based in
the key od the parent node, and the objectid of the loaded children.
This is a workaround since the entity based Nodes services can return nodes
with wrong keys.